Move worker enqueuing to session repository
