+++ app/src/main/java/io/rafaelribeiro/spendless/workers/UserSessionWorker.kt
@@ -3,11 +3,7 @@ package io.rafaelribeiro.spendless.workers
 import android.content.Context
 import android.util.Log
 import androidx.hilt.work.HiltWorker
-import androidx.work.Constraints
 import androidx.work.CoroutineWorker
-import androidx.work.ExistingWorkPolicy
-import androidx.work.OneTimeWorkRequestBuilder
-import androidx.work.WorkManager
 import androidx.work.WorkerParameters
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
@@ -15,7 +11,6 @@ import io.rafaelribeiro.spendless.domain.UserSessionRepository
 import io.rafaelribeiro.spendless.domain.user.UserSessionState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import java.util.concurrent.TimeUnit
 
 @HiltWorker
 class UserSessionWorker @AssistedInject constructor(
@@ -26,36 +21,7 @@ class UserSessionWorker @AssistedInject constructor(
 
     companion object {
         const val WORKER_TAG = "UserSessionWorker"
-        private const val WORK_NAME = "user_session_worker"
-
-        fun enqueue(context: Context, sessionExpiryDuration: Int) {
-            val workManager = WorkManager.getInstance(context)
-            val workInfo = workManager.getWorkInfosForUniqueWork(WORK_NAME).get()
-
-            if (workInfo.isNullOrEmpty() || workInfo.any { it.state.isFinished }) {
-                Log.i(WORKER_TAG, "Enqueueing Worker for $sessionExpiryDuration minutes...")
-                val constraints = Constraints.Builder()
-                    .setRequiresBatteryNotLow(false)
-                    .build()
-
-                val workRequest = OneTimeWorkRequestBuilder<UserSessionWorker>()
-                    .setInitialDelay(sessionExpiryDuration.toLong(), TimeUnit.MINUTES)
-                    .setConstraints(constraints)
-                    .build()
-
-                workManager.enqueueUniqueWork(
-                    WORK_NAME,
-                    ExistingWorkPolicy.REPLACE,
-                    workRequest
-                )
-            } else {
-                Log.i(WORKER_TAG, "Worker is already running or enqueued...")
-            }
-        }
-
-        fun cancel(context: Context) {
-            WorkManager.getInstance(context).cancelUniqueWork(WORK_NAME)
-        }
+        const val WORK_NAME = "user_session_worker"
     }
 
     override suspend fun doWork(): Result {
